Feign Exceptions Propagadas

+++ apiprodutos/src/main/kotlin/com/nadir/apiprodutos/entities/Entrega.kt
@@ -5,6 +5,6 @@ import java.math.BigDecimal
 data class Entrega(
     val idCliente: Long,
     val idProduto: Long,
-    val qtdItensComprados: Integer,
+    val qtdItensComprados: BigDecimal,
     val valorUnitarioDoItem: BigDecimal,
     val valorTotal: BigDecimal) {}
